move idea post to use idea post nodes instead of markdown nodes
