+++ gatsby-node.js
@@ -2,19 +2,13 @@ const _ = require("lodash");
 const fs = require("fs");
 const path = require("path");
 const { createFilePath } = require("gatsby-source-filesystem");
-const {
-    stringWithDefault,
-    resolveToArray,
-    resourceQuery,
-    ideaPostQuery,
-} = require("./gatsby/utils/gatsby-resolver-utils");
+const { createIdeaPostResolver } = require("./gatsby/resolvers/resolvers");
 const {
     RESOURCES_GATSBY_NODE_KEY,
     MARKDOWN_REMARK_GATSBY_NODE_KEY,
     TEMPLATE_KEY_TO_TYPE,
     ALLENITE_TEMPLATE_KEY,
     PROGRAM_TEMPLATE_KEY,
-    IDEA_POST_TEMPLATE_KEY,
 } = require("./gatsby/constants");
 
 const read = (p) => fs.readFileSync(path.join(__dirname, p), "utf8");
@@ -30,23 +24,7 @@ const DATA_ONLY_PAGES = [ALLENITE_TEMPLATE_KEY, PROGRAM_TEMPLATE_KEY];
 
 exports.createSchemaCustomization = ({ actions }) => {
     const { createTypes } = actions;
-    const typeDefs = [
-        `type PreliminaryFindings {
-            summary: String!
-            figures: [ImgWithCaption!]!
-        }
-
-        type ImgWithCaption @dontInfer {
-            type: String!
-            url: String
-            file: File @fileByRelativePath
-            caption: String
-        }
-
-        `,
-    ];
     createTypes(read("gatsby/schema/base.gql"));
-    createTypes(typeDefs);
 };
 
 /**
@@ -63,122 +41,7 @@ exports.createResolvers = ({ reporter, createResolvers }) => {
                 }),
             },
         },
-        Frontmatter: {
-            description: {
-                resolve: (source) =>
-                    stringWithDefault(
-                        source.description,
-                        "No description provided.",
-                    ),
-            },
-            authors: {
-                resolve: (source) => resolveToArray(source.authors),
-            },
-            title: {
-                resolve: (source) =>
-                    stringWithDefault(source.title, "No title provided."),
-            },
-            resources: {
-                resolve: async (source, _args, context) => {
-                    const names = resolveToArray(source.resources);
-                    const results = await Promise.all(
-                        names.map((name) =>
-                            context.nodeModel.findOne(resourceQuery(name)),
-                        ),
-                    );
-                    results.forEach((result, i) => {
-                        if (!result) {
-                            const msg = `Resource "${names[i]}" not found for idea "${source.title}". Check for typos and ensure the resource file exists with the correct templateKey.`;
-                            reporter.error(msg, new Error(msg));
-                        }
-                    });
-                    return results.filter(Boolean);
-                },
-            },
-            nextSteps: {
-                resolve: (source) => source.nextSteps ?? null,
-            },
-            program: {
-                resolve: (source) => resolveToArray(source.program),
-            },
-            preliminaryFindings: {
-                resolve: (source) => {
-                    const raw = source.preliminaryFindings;
-                    if (!raw || typeof raw !== "object") {
-                        return {
-                            summary: "",
-                            figures: [],
-                        };
-                    }
-                    return {
-                        summary: stringWithDefault(raw.summary, ""),
-                        figures: resolveToArray(raw.figures),
-                    };
-                },
-            },
-            relatedIdeas: {
-                resolve: async (source, _args, context) => {
-                    const names = resolveToArray(source.related_ideas);
-                    const results = await Promise.all(
-                        names.map((name) =>
-                            context.nodeModel.findOne(ideaPostQuery(name)),
-                        ),
-                    );
-                    results.forEach((result, i) => {
-                        if (!result) {
-                            const msg = `Idea post "${names[i]}" not found for idea "${source.title}". Check for typos and ensure the resource file exists with the correct templateKey.`;
-                            reporter.error(msg, new Error(msg));
-                        }
-                    });
-                    return results.filter(Boolean);
-                },
-            },
-        },
-        IdeaPost: {
-            relatedIdeas: {
-                resolve: async (source, _args, context) => {
-                    const names = resolveToArray(source.related_ideas);
-                    const results = await Promise.all(
-                        names.map((name) =>
-                            context.nodeModel.findOne(ideaPostQuery(name)),
-                        ),
-                    );
-                    results.forEach((result, i) => {
-                        if (!result) {
-                            const msg = `Idea post "${names[i]}" not found for idea "${source.title}". Check for typos and ensure the resource file exists with the correct templateKey.`;
-                            reporter.error(msg, new Error(msg));
-                        }
-                    });
-                    return results.filter(Boolean);
-                },
-            },
-            authors: {
-                resolve: (source) => resolveToArray(source.authors),
-            },
-            tags: {
-                resolve: (source) => resolveToArray(source.tags),
-            },
-            program: {
-                resolve: (source) => resolveToArray(source.program),
-            },
-            resources: {
-                resolve: async (source, _args, context) => {
-                    const names = resolveToArray(source.resources);
-                    const results = await Promise.all(
-                        names.map((name) =>
-                            context.nodeModel.findOne(resourceQuery(name)),
-                        ),
-                    );
-                    results.forEach((result, i) => {
-                        if (!result) {
-                            const msg = `Resource "${names[i]}" not found for idea "${source.title}". Check for typos and ensure the resource file exists with the correct templateKey.`;
-                            reporter.error(msg, new Error(msg));
-                        }
-                    });
-                    return results.filter(Boolean);
-                },
-            },
-        },
+        IdeaPost: createIdeaPostResolver(reporter),
     });
 };
 
@@ -190,19 +53,14 @@ exports.createResolvers = ({ reporter, createResolvers }) => {
 exports.createPages = ({ actions, graphql }) => {
     const { createPage } = actions;
 
-    // Typed node page creation is opt-in per template key.
-    // idea-post is excluded here until the template is migrated to query IdeaPost by id
-    // (currently it queries markdownRemark, so pages are created via markdownPages below).
-    const pagesToCreate = Object.keys(TEMPLATE_KEY_TO_TYPE).filter(
-        (templateKey) => templateKey !== IDEA_POST_TEMPLATE_KEY,
-    );
 
     // Create pages for any markdown files that are configured to have their
     // own node type (e.g. Resource) based on their templateKey.
-    const typedNodePages = pagesToCreate.map((templateKey) => {
-        const nodeKey = TEMPLATE_KEY_TO_TYPE[templateKey];
-        const allKeyString = `all${nodeKey}`;
-        return graphql(`
+    const typedNodePages = Object.keys(TEMPLATE_KEY_TO_TYPE).map(
+        (templateKey) => {
+            const nodeKey = TEMPLATE_KEY_TO_TYPE[templateKey];
+            const allKeyString = `all${nodeKey}`;
+            return graphql(`
         {
             ${allKeyString} {
                 nodes {
@@ -212,20 +70,23 @@ exports.createPages = ({ actions, graphql }) => {
             }
         }
     `).then((result) => {
-            if (result.errors) {
-                result.errors.forEach((e) => console.error(e.toString()));
-                return Promise.reject(result.errors);
-            }
+                if (result.errors) {
+                    result.errors.forEach((e) => console.error(e.toString()));
+                    return Promise.reject(result.errors);
+                }
 
-            result.data[allKeyString].nodes.forEach((node) => {
-                createPage({
-                    path: node.slug,
-                    component: path.resolve(`src/templates/${templateKey}.tsx`),
-                    context: { id: node.id },
+                result.data[allKeyString].nodes.forEach((node) => {
+                    createPage({
+                        path: node.slug,
+                        component: path.resolve(
+                            `src/templates/${templateKey}.tsx`,
+                        ),
+                        context: { id: node.id },
+                    });
                 });
             });
-        });
-    });
+        },
+    );
 
     /**
      * We make pages from all markdown files that are consumed by gatsby-transformer-remark,
@@ -265,7 +126,7 @@ exports.createPages = ({ actions, graphql }) => {
             // Skip creating pages for data-only pages (software, dataset, etc.)
             if (
                 DATA_ONLY_PAGES.includes(templateKey) ||
-                pagesToCreate.includes(templateKey)
+                Object.keys(TEMPLATE_KEY_TO_TYPE).includes(templateKey)
             ) {
                 return;
             }
