+++ src/types/index.ts
@@ -13,18 +13,16 @@ import { IGatsbyImageData } from "gatsby-plugin-image";
  * After updating schema customization, run `gatsby develop` to regenerate types.
  */
 export type IdeaPostQuery = Queries.IdeaPostByIDQuery;
-export type IdeaPostNode = NonNullable<IdeaPostQuery["markdownRemark"]>;
-export type IdeaFrontmatter = NonNullable<IdeaPostNode["frontmatter"]>;
-export type IdeaFields = NonNullable<IdeaPostNode["fields"]>;
+export type IdeaPostNode = NonNullable<IdeaPostQuery["ideaPost"]>;
 
 export type IdeasForTags =
     Queries.AllIdeasForTagsQuery["allMarkdownRemark"]["edges"];
 
 export type PreliminaryFindings = NonNullable<
-    IdeaFrontmatter["preliminaryFindings"]
+    IdeaPostNode["preliminaryFindings"]
 >;
 export type Figure = NonNullable<
-    NonNullable<IdeaFrontmatter["preliminaryFindings"]>["figures"]
+    NonNullable<IdeaPostNode["preliminaryFindings"]>["figures"]
 >[number];
 
 export type ResourceTemplateQuery = Queries.ResourcesByIdQuery;
